+++ mono_repo/lib/src/package_config.dart
@@ -24,6 +24,15 @@ enum PackageFlavor {
 }
 
 extension PackageFlavorExtension on PackageFlavor {
+  String get pubCommand {
+    switch (this) {
+      case PackageFlavor.flutter:
+        return 'flutter pub pub';
+      case PackageFlavor.dart:
+        return 'dart pub';
+    }
+  }
+
   String get prettyName {
     switch (this) {
       case PackageFlavor.dart:
@@ -45,7 +54,7 @@ extension PackageFlavorExtension on PackageFlavor {
 
       case PackageFlavor.flutter:
         return {
-          'uses': 'subosito/flutter-action@v1.5.3',
+          'uses': 'subosito/flutter-action@v2.0.4',
           'with': {
             'channel': sdkVersion,
           }